Removed the checks for asymtotic rays from the inner loop of the ray caster (these have never triggered during development).
Moved those checks to the build stage (of the lookup tables), so the application now asserts if we ever create a possible asymtotic ray. (at which point, ad the run-time handling back, I guess).

Optimized the computation of the current cell in the ray caster - using integer shifts instead of divisions.

Created a constexpr Utils::log2-function to calculate the amount we need to shift (eg: log base 2 of CELL_WIDTH and CELL_HEIGHT). Of course this only works with base2 friendly cell dimensions. :)